moved test to more relevant test file
